+++ src/types/helpers.ts
@@ -65,3 +65,12 @@ export function convertToType(Target: any, data?: object): object | undefined {
 
   return Object.assign(new Target(), data);
 }
+
+export function getEnumValuesMap<T extends object>(enumObject: T) {
+  const enumKeys = Object.keys(enumObject).filter(key => isNaN(parseInt(key, 10)));
+  const enumMap = enumKeys.reduce<any>((map, key) => {
+    map[key] = enumObject[key as keyof typeof enumObject];
+    return map;
+  }, {});
+  return enumMap;
+}