+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog and release notes
 
+## Unreleased
+### Features
+- add base support for GraphQL enums using TypeScript enums
+
 ## v0.7.0
 ### Features
 - add authorization feature - `@Authorized` decorator and `authChecker` function in schema options ([see docs](https://github.com/19majkel94/type-graphql/blob/master/docs/authorization.md))